Move LUTs in Hsi before register outputs
If the inverted registers are assigned to the outputs the
synthesis will result in a circuit that has inverters placed
between the pin and the final register. Since this has a
negative effect on the propagation time of the signal out
of the FPGA, move inverters in front of registers.