Shrink SpinalTap internal Apb bus
With the bus set to 9 bit and the address space
per module being 256 byte, the last bit of the
module address will be passed on to the modules.
Those in turn will then not react on read/write
via the bus since the register addresses will not
match.
E.g. 0x43c00108 becomes 0x108 after the interconnect
but the module will only have a register at
0x008 (because of the 256byte address space)